move 'whisper_echo/examples/whisper_params*' to 'whisper_echo/examples/whisper_params*'
