+++ whisper_echo/CMakeLists.txt
@@ -1,1 +1,3 @@
 add_subdirectory(examples)
+
+add_library(whisper_params OBJECT whisper_params.cc)
