+++ whisper_echo/examples/main.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 
 #include "whisper.h"
-#include "whisper_echo/examples/whisper_params.h"
+#include "whisper_echo/whisper_params.h"
 
 int main(int argc, char** argv) {
   whisper::WhisperParams params(argc, argv);
